# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnAI - AI-Powered Adaptive Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all AI tutor conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new AI tutor conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/feynman-check:
    post:
      operationId: postOpenaiFeynmanCheck
      tags: [openai]
      summary: Evaluate a Feynman-style explanation for clarity and gaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeynmanCheckInput"
      responses:
        "200":
          description: Score and improvement issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeynmanCheckOutput"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiImageOutput"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── NOTE ──────────────────────────────────────────────────────────────────
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
        courseId:
          type: integer
        stepId:
          type: integer
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
        mode:
          type: string
          enum: [socratic, direct, beginner, exam_coach, deep_theory, practical, error_diagnosis, feynman, math, memory_coach]
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
...
    FeynmanCheckInput:
      type: object
      properties:
        explanation:
          type: string
          minLength: 10
        concept:
          type: string
      required:
        - explanation
    FeynmanCheckOutput:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        issues:
          type: array
          items:
            type: string
        isDemo:
          type: boolean
      required:
        - score
        - issues
    OpenaiImageInput:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
      required:
        - prompt
    OpenaiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...